[Bronze V] Title: 이상한 기호, Time: 176 ms, Memory: 17652 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,7 +6,6 @@ public class Main {
 
         int a = scan.nextInt();
         int b = scan.nextInt();
-        
-        System.out.print((a+b)*(a-b));
+        System.out.print((long)(a+b)*(a-b));
     }
 }
